Initial unity/nux atk module implementation
This commit includes the skeleton implementation for most of the ATK objects,
that provides the accessibility support for the related nux and unity objects.

Those objects doesn't provide too much functionality, but with them you can
obtain the tree object hierarchy of the Launcher from a AT tool. This objects
will be used as a base to work on. The next step will be fill them with
functionality.

This commit provides:

  * ATK objects for the nux objects Object, Area, View, Layout and BaseWindow.

  * ATK objects for the unity objects Launcher and LauncherIcon

  * As we don't want to add the accessibility tips directly on Nux, it is
    provided a factory like method, that maintains the reference of each
    accessible object on a hash table.

  * This factory method is based on the Object systems provided by Nux.

  * For this reason we needed to add Launcher and LauncherIcon on this
    object system (anyway, this was likely a error)

  * For the moment the Launcher window is added manually to the root. In
    the short term this will be improved by a signal-callback system.

  * We also provide some basic functionality on some objects:
    * The container AtkObject methods on all the objects (to get the hierarchy)
    * Screen position on Area
    * ATK object for LauncherIcon expose the LauncherIcon tooltip as the
      accessible object name

(bzr r750.1.2)