Fading title: consider padding for long titles, don't fade the title too early
The title should be faded just few pixel (currently I've
increased the fading space to 30) before the end of the
Panel Menuview (so before the indicators).
The previous implementation was wrong, also due to the fact
that for title slightly longer than the panel space no
cut was applied.

(bzr r842.1.1)