I have added the quicklist item "Keep in Launcher" also for the device icons!
In addition i have added a ccsm option "Show Devices" by the which we can choose to display all devices, only mounted devices and no one.

The "keep in launcher" works only if the option "Show Devices" is set to "Only Mountes" and if the device is not an automounted one!

(bzr r968.3.1)