unityshell: Send SecondaryActivate signal on indicator middle click
When an indicator item is middle-clicked the SecondaryActivate
method is called. The real click position is exported as requested
by the StatusNotifierItem specifications, and the event is
emitted only if the user middle-click and release over an indicator.
(so the mouse click can be canceled releasing the mouse click
outside the indicator entry space).

(bzr r1283.2.2)