Added PanelIndicatorsView to move to the priority-based system for ordering entries
Removed PanelIndicatorObject and replaced with PanelIndicatorsView that
basically is a container of indicators which shows them following the priority
of the given entries.

So, now PanelView just contains two indicator children: menu_view and indicators
They get updated when a new entry is added and uses the entries priority
to order them. Thanks to this now the indicators can be also in a mixed order
and the entries aren't anymore ordered just per-IndicatorObject.

(bzr r1351.2.4)