Fixing again bug #820293 (using a cleaner way) and #835637
So, now the panel view is hidden also when leaving the panel from the
left edge of the panel.

Bug #835731 can't fixed in PanelMenuView directly, since the it doesn't
get events when the pointer is positioned on the first screen column of pixels
(also if it uses that area, actually).

(bzr r1468.1.1)