Don't consider unmapped and unviewable windows in IsWindowObscured
An unmapped or not viewable window can't for sure obscure another window.
So, when trying to collect the maximized Windows in PanelMenuWindow
(to restore them grabbing from the panel), the unmapped windows
(like banshee in the sound menu or deluge reduced to its indicator)
must not be considered or the dragging wouldn't work.

This should fix bugs like #781215 and #774121

(bzr r1468.1.16)