panel-service: Don't remove all release events outside the menu area on first click
Unity panel service, after opening a menu was ignoring all the mouse release events
to avoid to close the just opened menu. But this was incorrect and caused
bug #844309 so it was impossible click on a menu, continue clicking over it
and releasing the mouse outside its area (to close it).

So now we only ingore release events that are above the menu, so if
a mouse button is released above the menu (and so in its title in the panel)
the release event is ignored and so the menu doesn't close, otherwise
the menu closes.

(bzr r1468.1.18)