Panel-service: cache the entry geometries and use them for parsing mouse events
When a client calls the SyncGeometry method the panel-service caches
the geometries for the changed entries and then we use them in the
event filter function to check if a release event has been done
over an indicator entry name or not.

This fixes for good bug #844309 and improves the menus handling with
a single-click: when a click is released over a non-previously selected
menu title, then the release event is ignored, otherwise we use it
to close the menus.

(bzr r1468.1.19)