PanelView: track mouse position only when it has been changed
We cache the tracked mouse position to avoid to check for changed
position every few milliseconds without any need.

This commit fixes bug #834065 since with the previous implementation,
when changing the active menu using the keyboard, the mouse position was
always checked for opening a menu. Now it's relevant only if it has been
changed.

(bzr r1468.1.27)