StaticCairoText: fix RTL width calculation.
Pango will calculate logRect.x as the left-most point of the string, and
start rendering the string at (0,0).  For right-to-left languages
that means that logRect.x will always be <= -logRect.width.

StaticCairoText does not use logRect.x to offset the text rendering,
so just ignore it in the extent calculation.

Fixes LP: #807921

(bzr r1738.10.4)