PanelMenuView: Use the new Animator Start to use different fade timings
When a new application is shown, to use different fade-in/out timings,
we use the new Start function that allows to define a per-animation
duration, this makes the code much more clean.
By the way we need a boolean variable to track when we need to fade-out
slowly.

(bzr r1738.7.11)