BamfLauncherIcon: manage user-invisible windows
Windows that have the skip-taskbar flag shouldn't be managed by
unity unless there's another user-visible window of the given application
mapped or the application is pinned.

Also, when performing a click over a launcher icon, if it has currently
opened only non-visible windows, the application is launched, otherwise
we give priority to the user visible windows bringing to focus just them
(this doesn't happen in the case of the Alt+Tab).

When an application with user-invisible windows is closed, we show it
on the launcher only if that is pinned.

(bzr r1815.1.1)