PanelMenuView: First working stub of integrated menus
The logic is all here, we need to improve the design.
However, now when integrated menus are enabled, when adding an entry to
the panel, we just consider it as the integrated-menu and we create a new
PanelIndicatorAppmenuView that will handle that.
The view is also updated everytime the panel title or the active window
changes.

(bzr r1858.9.75)