PanelTitlebarGrabAreaView: redesigned, to getter grabbed event using a timeout
If the grab area get a mouse-down event, if we get a mouse-up event before a defined timeout
(set to 120ms now) occurs, then we consider that a click. Otherwise we send the grab_started event

This allows to improve the way we handle the grab event in the panel, and allow to avoid to
change the mouse cursor if not needed.

(bzr r1858.9.103)