PanelIndicatorsView: make each menu entry to use the most space it really can
When setting the maximium width for the entries, we shouold care of the
current sizes. So basically, if an entry doesn't need to be resized,
we can just take its width and remove it from the maximum space we have,
then the space that the other entries can use, is computed using the
remaining space.

This improves a lot the entries collapsing, and make sure that we always
use all the horizontal space that we have.

(bzr r1858.17.13)