ubus: Ensure the GType system has been initialized before using any GTypes.
The unity plugin doesn't explicitly call g_type_init(), although it uses a
GObject for UBusServer. This currently works because the ccp plugin, which
gets loaded before the unity plugin, calls g_type_init(), but this is fragile.
If ccp is not used, g_type_init() is not called by the time UBusServer gets
initialized, leading to a crash. This patch ensures that the GType system
has been initialized before using UBusServer (it is safe to call g_type_init()
multiple times).

(bzr r2120.6.1)