unityshell gestures: Use hit-test to find window affected by 3-touch drag
Instead of climbing up the window hierarchy with XQueryTree. That will improve
performance.

The root cause of lp:940612 is that after the re-architecture of the UTouch stack
the child_window_id event attribute is no longer being filled. Even though
UTouch stack should have it fixed, the change made by this patch is still
an improvement on its own right. Besides, in the new architecture, UTouch might
have trouble filling this attribute in the same way as before.

Fixes LP: #940612

(bzr r2198.1.1)