GLibSource: add CallBackData struct, since the DestroyCallback can be called later
We add a new struct that is used to pass the Source pointer to the callbacks
since the DestroyCallback can be called even after that the wrapper
has been destroyed. This would cause a seg-fault.

At this point the Manager doesn't need to check when we remove a source.

(bzr r2364.2.25)