SoftwareCenterLauncherIcon: remove the icon size parameter from the animation. Change logic.
We assume now that when a SoftwareCenterLauncher icon has been added to the launcher, is already
on the model, but it is not visible. So at this point we have only to animate ourself to 
the position that was designed (and saved on the icon Center).

When the animation is over, remove any extra reference to icon textures and launcher.
Also make sure we're visible.

(bzr r2529.14.47)