Add logind alternatives to ConsoleKit calls
ConsoleKit is being deprecated, so add support for logind for logout, shutdown,
and reboot.

With ConsoleKit, sessions get $XDG_SESSION_COOKIE set, while with logind they
get $XDG_SESSION_ID. So use that to determine which one to talk to until we
drop support for ConsoleKit.

Fixes LP: #1155021

(bzr r3226.3.1)