UnityScreen: remove the useless and expensive gl{Push,Pop}Attrib calls
For some reasons this code was copied by the opengl plugin as a workaround
to fix the state of our screen after that nux has drawn.
Actually this is not needed, the only thing we really need to do is to fix the
current Viewport, because nux seems to leave it in a bad state which would
lead to flickering menus, fullscreen windows, tooltip and missing windows
thumbnails in switcher.

Thanks to Sam Spilsbury for his precious support.

(bzr r3592.1.1)