BackgroundEffectHelper: keep a cached version of blur geometries
Now the blurred geometries are saved in a list that gets updated only
when they change, this allows to reduce a lot the cost of GetBlurGeometries
that now only returns a const reference of the actual vector.

Improved also HasEnabledHelpers (no helper can be registered if it's not enabled).

(bzr r3347.5.20)