SwitcherView: only redraw the view if an icon changed since the last progress iteration
We don't want the switcher to be changed unless its geometry or an icon changes.
Also we don't want to redraw the switcher multiple times if only a not-animated parameter
of an icon (such as a pip or the selection glow) has changed.

As bonus, reduce lots of list copies.

(bzr r3347.6.9)