DecoratedWindow: add last_title string to backup the title name in case the window is undecorated
This is mostly for maximized windows, as they do have the title also if the current decoration::Title has
been deleted with the top layout. In this case we need to keep a copy of the text around until we've
not get decorated again. We're avoiding to duplicate any string content, though.

(bzr r3566.5.242)