IndicatorEntry: cleanup the API, add parents support, improve testing
Each Entry might have one or more (it might happen on multi-monitor) parent entries that will
match a child geometry, active state and show-now state.

So, when we add a parent we update its values to match the child, when a relevant child value changes
we inform the parents about it.

(bzr r3566.5.282)