LockScreenController: wait the primary shield to get the grab before setting the session locked
If that fails (and if the grab has never been gained during this locking session),
it's up to the Controller to decide whether the lock request has to be ignored.
So, if the primary shield fails to grab, before that the screen has ever been locked,
then we have to give up. Otherwise if it fails (but we were already locked), it's safe
to assume that this failure is due to the menus that are currently controlling the grab.

Also setup the primary shield signal connections in a shared place and use a
connection::Manager to keep track of them.

Fixes LP: #1368427

(bzr r3871.3.3)