CompizUtils: add WindowDecorationElements that returns a bitset describing the elements available
This function will allow to check which decoration elements are supported by a window
by using just a single call, that will check which elements should be added when decorating
the window. By returning a single unsigned that is a bitwise composition of values, we can
just use a single call to verify which elements are supported.

(bzr r3806.12.23)