ApplicationManager: rely on windows monitor property changes for updating Pips
Removing the expensive calls that rely on Compiz window moved events which
are fired for every window, also if outside the scope of an Icon, so we don't
have to filter these signals out, but we can just get the monitor.changed signal
and depend on that.
This fixes the launcher pips not to be properly initialized in multi-monitor environments.

Fixes LP: #1027191

(bzr r3983.3.10)