DevicesSettingsImp: dont't feed garbage to gsettings
Use a const reference for the iteration variable in the loop over
_blacklist so that the pointers returned from c_str() are still valid
when feeding them to g_settings_set_strv().

In the best case, this only wrote garbage to the "blacklist" key (making
that setting useless). In the worst case, it led to a crash because the
garbage was invalid utf8, which gsettings refuses to accept.

Fixes LP: #1495173

(bzr r4008.2.1)